# webgpu: Optimize maxpool with groupsize(4, 4, 4)