--- a/tfjs-backend-webgpu/src/benchmark_ops_test.ts
+++ b/tfjs-backend-webgpu/src/benchmark_ops_test.ts
@@ -135,4 +135,11 @@ describeWebGPU('Ops benchmarks', () => {
 
     await time(() => tf.depthwiseConv2d(x, w, 1, 'valid'));
   });
+
+  it('maxPool', async () => {
+    const x = tf.randomNormal<tf.Rank.R4>([1, 131, 131, 64]);
+
+    await time(() => tf.maxPool(x, 2, 1, 'same'));
+  });
+
 });